requirements.txt
streamlit
pandas
matplotlib
arabic_reshaper
python-bidi
openpyxl